cleaner version of initialBackstack
