+++ app/src/main/java/io/github/wiiznokes/gitnote/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.lifecycle.viewmodel.compose.viewModel
 import dev.olshevski.navigation.reimagined.AnimatedNavHost
 import dev.olshevski.navigation.reimagined.NavBackHandler
@@ -15,11 +15,8 @@ import dev.olshevski.navigation.reimagined.navigate
 import dev.olshevski.navigation.reimagined.popAll
 import dev.olshevski.navigation.reimagined.popUpTo
 import dev.olshevski.navigation.reimagined.rememberNavController
-import io.github.wiiznokes.gitnote.MyApp.Companion.appModule
-import io.github.wiiznokes.gitnote.helper.NoteSaver
 import io.github.wiiznokes.gitnote.ui.destination.AppDestination
 import io.github.wiiznokes.gitnote.ui.destination.Destination
-import io.github.wiiznokes.gitnote.ui.destination.EditParams
 import io.github.wiiznokes.gitnote.ui.destination.SetupDestination
 import io.github.wiiznokes.gitnote.ui.screen.app.AppScreen
 import io.github.wiiznokes.gitnote.ui.screen.setup.SetupNav
@@ -58,31 +55,11 @@ class MainActivity : ComponentActivity() {
                 dynamicColor = dynamicColor
             ) {
 
-                val startDestination: Destination = remember {
+                val startDestination: Destination = rememberSaveable {
                     if (runBlocking { vm.tryInit() }) {
-                        if (NoteSaver.isEditUnsaved()) {
-                            val saveInfo = NoteSaver.getSaveState()
-                            if (saveInfo == null) {
-                                Log.d(TAG, "can't retrieve the last saved note state")
-                                Destination.App(AppDestination.Grid)
-                            } else {
-                                Log.d(TAG, "launch as EDIT_IS_UNSAVED")
-                                Destination.App(
-                                    AppDestination.Edit(
-                                        EditParams.Saved(
-                                            note = saveInfo.previousNote,
-                                            editType = saveInfo.editType,
-                                            name = saveInfo.name,
-                                            content = saveInfo.content
-                                        )
-                                    )
-                                )
-                            }
-
-                        } else Destination.App(AppDestination.Grid)
+                        Destination.App(AppDestination.Grid)
                     } else Destination.Setup(SetupDestination.Main)
                 }
-
 
                 val navController =
                     rememberNavController(startDestination = startDestination)
